move tests to outside of landsat folder
