+++ tests/test_search.py
@@ -3,15 +3,9 @@
 
 """Tests for search"""
 
-import os
-import sys
 import unittest
 
-try:
-    from landsat.search import Search
-except ImportError:
-    sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '../landsat')))
-    from landsat.search import Search
+from landsat.search import Search
 
 
 class TestSearchHelper(unittest.TestCase):
